ml-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Machine Learning and NLP
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Sentiment Analysis Models
vaderSentiment==3.3.2
textblob==0.17.1

# Topic Modeling
bertopic==0.15.0
umap-learn==0.5.4
hdbscan==0.8.33

# Text Processing
nltk==3.8.1
spacy==3.7.2
regex==2023.10.3

# HTTP and API clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Database and caching
motor==3.3.2
redis==5.0.1
pymongo==4.6.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Image processing (for future features)
Pillow==10.1.0

# Async support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
